factorized the architecture
