+++ hungarian_net/torch_modules/attention_layer.py
@@ -1,0 +1,36 @@
+from functools import partial
+from typing import Any, Dict, Tuple
+
+import lightning as L
+import torch
+import torch.nn as nn
+import torchmetrics
+from lightning.pytorch.utilities.types import STEP_OUTPUT
+from sklearn.metrics import f1_score
+from torch import optim
+from torchmetrics import MetricCollection
+
+class AttentionLayer(nn.Module):
+    def __init__(self, in_channels, out_channels, key_channels):
+        super(AttentionLayer, self).__init__()
+        self.conv_Q = nn.Conv1d(in_channels, key_channels, kernel_size=1, bias=False)
+        self.conv_K = nn.Conv1d(in_channels, key_channels, kernel_size=1, bias=False)
+        self.conv_V = nn.Conv1d(in_channels, out_channels, kernel_size=1, bias=False)
+
+    def forward(self, x):
+        Q = self.conv_Q(x)
+        K = self.conv_K(x)
+        V = self.conv_V(x)
+        A = Q.permute(0, 2, 1).matmul(K).softmax(2)
+        x = A.matmul(V.permute(0, 2, 1)).permute(0, 2, 1)
+        return x
+
+    def __repr__(self):
+        return (
+            self._get_name()
+            + "(in_channels={}, out_channels={}, key_channels={})".format(
+                self.conv_Q.in_channels,
+                self.conv_V.out_channels,
+                self.conv_K.out_channels,
+            )
+        )
